+++ CHANGELOG.md
@@ -4,7 +4,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2021-03-16
+
+### Added
+- Expose lifecycle hook events. This feature allows sending VWO data to third party integrations.
+
+### Changed
+
+- Introduce withIntegrations method in launch API to enable receiving hooks for the third party integrations.
+
+```java
+     public IntegrationEventListener getEventData() {
+        return new IntegrationEventListener() {
+          @Override
+          public void onEvent(Map<String, Object> map) {};
+        }
+    };
+
+    String settingsFile = VWOHelper.getSettingsFile(accountId, sdkKey);
+    VWO vwoInstance = VWO.launch(settingsFile).withIntegrations(getEventData()).build();
+```
 ## [1.11.0] - 2021-03-10
+
 ### Changed
 #### 1. Event Batching